third_party/WebKit/LayoutTests/web-animations-api/startTime-expected.txt

This is a testharness.js-based test.
FAIL startTime of a newly created (idle) animation is unresolved Animation is not defined
FAIL startTime of a play-pending animation is unresolved Animation is not defined
FAIL startTime of a pause-pending animation is unresolved Animation is not defined
PASS startTime of a play-pending animation created using Element.animate shortcut is unresolved 
PASS startTime is resolved when running 
PASS startTime and currentTime are unresolved when animation is cancelled 
Harness: the test ran to completion.
